+++ CHANGELOG.md
@@ -6,10 +6,11 @@
 
 * **NEW** 🎉
   * Added option to open results rows/values in editor. Issue #140.
-  * Added SSL support for postgres viw `pgOptions` connection setting. Issue #141
+  * Added SSL support for postgres via `pgOptions` connection setting. Issue #141
   * Deprecating `dialectOptions` in favor of `[dialect]Options` settings.
 
 * **Fixes**
+  * Fixed dollar quoting on formatter. Issue #142
 
 ### v0.17.5